+++ zojae031/Project/app/src/main/java/study/architecture/coinjob/CoinDataAdapter.kt
@@ -1,4 +1,4 @@
-package study.architecture.ui.coinjob
+package study.architecture.coinjob
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
