+++ zojae031/Project/app/src/main/java/study/architecture/coinjob/CoinViewModel.kt
@@ -1,4 +1,4 @@
-package study.architecture.ui.coinjob
+package study.architecture.coinjob
 
 import android.util.Log
 import androidx.databinding.ObservableField
