+++ zojae031/Project/app/src/main/java/study/architecture/data/datasource/local/CoinDatabase.kt
@@ -1,4 +1,4 @@
-package study.architecture.data.local
+package study.architecture.data.datasource.local
 
 import android.content.Context
 import androidx.room.Database
@@ -6,8 +6,8 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 import study.architecture.data.entity.Market
 import study.architecture.data.entity.Ticker
-import study.architecture.data.local.dao.MarketDao
-import study.architecture.data.local.dao.TickerDao
+import study.architecture.data.datasource.local.dao.MarketDao
+import study.architecture.data.datasource.local.dao.TickerDao
 
 @Database(entities = [Market::class, Ticker::class], version = 1)
 abstract class CoinDatabase : RoomDatabase() {
