+++ zojae031/Project/app/src/main/java/study/architecture/data/datasource/local/LocalDataSource.kt
@@ -1,10 +1,13 @@
-package study.architecture.data.remote
+package study.architecture.data.datasource.local
 
 import io.reactivex.Single
 import study.architecture.data.entity.Market
 import study.architecture.data.entity.Ticker
 
-interface RemoteDataSource {
+interface LocalDataSource {
     fun getMarkets(): Single<List<Market>>
     fun getTickers(markets: String): Single<MutableList<Ticker>>
+    fun insertMarket(market: Market)
+    fun insertTicker(ticker: Ticker)
+
 }
