+++ zojae031/Project/app/src/main/java/study/architecture/data/datasource/local/LocalDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package study.architecture.data.local
+package study.architecture.data.datasource.local
 
 import io.reactivex.Single
 import io.reactivex.schedulers.Schedulers
