+++ zojae031/Project/app/src/main/java/study/architecture/data/datasource/local/dao/MarketDao.kt
@@ -1,4 +1,4 @@
-package study.architecture.data.local.dao
+package study.architecture.data.datasource.local.dao
 
 import androidx.room.Dao
 import androidx.room.Query
