+++ zojae031/Project/app/src/main/java/study/architecture/data/datasource/remote/RemoteDataSource.kt
@@ -1,13 +1,10 @@
-package study.architecture.data.local
+package study.architecture.data.datasource.remote
 
 import io.reactivex.Single
 import study.architecture.data.entity.Market
 import study.architecture.data.entity.Ticker
 
-interface LocalDataSource {
+interface RemoteDataSource {
     fun getMarkets(): Single<List<Market>>
     fun getTickers(markets: String): Single<MutableList<Ticker>>
-    fun insertMarket(market: Market)
-    fun insertTicker(ticker: Ticker)
-
 }
