+++ zojae031/Project/app/src/main/java/study/architecture/data/datasource/remote/RemoteDataSourceImpl.kt
@@ -1,4 +1,4 @@
-package study.architecture.data.remote
+package study.architecture.data.datasource.remote
 
 import io.reactivex.Single
 import io.reactivex.schedulers.Schedulers
