+++ zojae031/Project/app/src/main/java/study/architecture/mainjob/MainActivity.kt
@@ -1,4 +1,4 @@
-package study.architecture.ui.mainjob
+package study.architecture.mainjob
 
 import android.os.Bundle
 import android.view.Menu
