+++ zojae031/Project/app/src/main/java/study/architecture/mainjob/MainPageAdapter.kt
@@ -1,10 +1,10 @@
-package study.architecture.ui.mainjob
+package study.architecture.mainjob
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentPagerAdapter
-import study.architecture.ui.coinjob.CoinFragment
+import study.architecture.coinjob.CoinFragment
 
 /**
  * [FragmentPagerAdapter] vs [FragmentStatePagerAdapter]
