+++ zojae031/Project/app/src/main/java/study/architecture/mainjob/MainViewModel.kt
@@ -1,0 +1,7 @@
+package study.architecture.mainjob
+
+import study.architecture.coinjob.CoinFragment
+
+class MainViewModel {
+    val pageLimit: Int = CoinFragment.FragIndex.values().size
+}
