--- zojae031/Project/app/src/main/java/study/architecture/ui/mainjob/MainViewModel.kt
@@ -1,7 +1,0 @@
-package study.architecture.ui.mainjob
-
-import study.architecture.ui.coinjob.CoinFragment
-
-class MainViewModel {
-    val pageLimit: Int = CoinFragment.FragIndex.values().size
-}
