+++ zojae031/Project/app/src/main/java/study/architecture/util/AdapterUtil.kt
@@ -3,7 +3,7 @@ package study.architecture.util
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.RecyclerView
 import study.architecture.data.entity.ProcessingTicker
-import study.architecture.ui.coinjob.CoinDataAdapter
+import study.architecture.coinjob.CoinDataAdapter
 
 @BindingAdapter("replaceAll")
 fun RecyclerView.replaceAll(list: List<ProcessingTicker>?) {
