+++ zojae031/Project/app/src/main/res/layout/activity_main.xml
@@ -15,7 +15,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.mainjob.MainActivity">
+        tools:context=".mainjob.MainActivity">
 
         <LinearLayout
             android:id="@+id/data_wrapper"
